search bar updates the filter title <-> vice versa
last commit, just added a few  stuff with the search and filter connection ->> modifying in real time, meaning that if I change the input in the search bar or in the side panel, it will automatically change my other value in the other component, without pressing Enter. small bug with double filtering, will fix it!